backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
email-validator==2.3.0
pydantic==2.11.9
